Update
+fixed movement issues
+Hinzufügen eines MausListeners
+mehrere int -->float
+Test Mobs am Spawn
+schießen eines Arrow objektes mit der linke Maus Taste

+++ src/entitys/Player.java
@@ -3,7 +3,7 @@ package entitys;
 public class Player extends MobTemplate {
 	public int AtkCooldown = 0;
 	public int HitCooldown = 0;
-	public Player(int hoehe, int breite, int posX, int posY, int dx, int dy, int speed, int SpawnX, int SpawnY,	int damage, int Hp) {
+	public Player(int hoehe, int breite, int posX, int posY, int dx, int dy, float speed, int SpawnX, int SpawnY,	int damage, int Hp) {
 		super(hoehe, breite, posX, posY, dx, dy, speed, SpawnX, SpawnY, damage, Hp);
 	}
