+++ src/gameObject/Rechteck.java
@@ -3,11 +3,11 @@ package gameObject;
 public class Rechteck {
 	public int hoehe;
 	public int breite;
-	public int posX;
-	public int posY;
+	public float posX;
+	public float posY;
 
 	
-	public Rechteck(int hoehe,int breite, int posX, int posY) {
+	public Rechteck(int hoehe,int breite, float posX, float posY) {
 		this.hoehe = hoehe;
 		this.breite = breite;
 		this.posX = posX;
